Issue #11163: Split InputHiddenField4 into focused test inputs
